// src/Shared/scripts/Enums/APICommand.ts
export enum APICommandFlag {
    Unknown = 0,
    PresentationDetails = 1,
    NavigateLinks = 2,
    ToggleRawValues = 3,
    NavigateBack = 4,
    NavigateForward = 5,
    NavigateUp = 6,
    GetStateOfScUiProxy = 7,
    AddContentEditorToDesktopAsync = 8,
    PublischActiveCE = 9,
    ToggleCompactCss =10,
    CEGoSelected=11,
    SetStateOfSitecoreWindowAsync = 12,
    OpenContentEditor = 13,
    AdminB = 14
}